+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * fix linkify urls in command line (enable link in cmd)
 * fix two bugs with echo without newline (when prompt have 0 width and lines have different length)
 * fix applying glow shadow to transparent text
+* fix async responsive greetings
+* fix multiple echo of async functions and promises mixed with strings
 
 ## 2.29.1
 ### Bugfix